Create money
As this is the first time ever working with c++ and visual studio, I used a bit of code I found on the internet(http://www.cprogramming.com/challenge.html), and modified to in the sole purpose of testing the waters. In it original form, an int was converted and outputted to words. I modified it to take a double and print it out in dollars and cents.